moved PackagerFactory to packagers
